attack_mitigation_tool/requirements.txt
flask==2.0.1
flask-wtf==0.15.1
pymysql==1.0.2
python-dotenv==0.19.0
werkzeug==2.0.1
wtforms==2.3.3
cryptography==3.4.8